+++ CHANGELOG.md
@@ -23,6 +23,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - tracing: improve performance of `TraceId::generate()`.
 - dumping: remove unstable `Timestamp`.
 
+### Fixed
+- core/request: avoid all-request to multiple groups hanging in some cases ([#127]).
+
 [#52]: https://github.com/elfo-rs/elfo/issues/52
 [#127]: https://github.com/elfo-rs/elfo/pull/127
 [#128]: https://github.com/elfo-rs/elfo/pull/128